refactor(lib): extract shared Flock portal constants and json utils
BASE_URL, USER_AGENT, FAILED_FILE, load_json, save_json, and dedupe are
now in lib.py. flock_transparency.py and the upcoming slug_probe.py both
use them. save_json also mkdirs its parent, which is harmless for the
existing callers and saves the new caller from duplicating the guard.
